*** Added offline processing for sessions.  Now a user can load a .SAZ file and process the data offline in Fiddler/Watcher.   % Fixed progress bar to move incrementally.
TODO: I removed the system/menu options from the ProgressDialog by setting ControlBox = False. This removes the ability for a user to close the ProgressDialog.
It's not ideal but it prevents the control from being Disposed prematurely.  Should later enable the ControlBox so people can close the ProgressDialog if they want.

--HG--
branch : 1.5